+++ src/index.js
@@ -1,10 +1,10 @@
-const { serverConfig, loggerConfig } = require('./config');
+const { serverConfig, loggerConfig, keycloakRetrieveConfig } = require('./config');
 const Server = require('./server/server');
 
 new Server(serverConfig)
 	.configureHelmet()
 	.configureLogging(loggerConfig)
-	.configureKeycloakRetrival()
+	.configureKeycloakRetrival(keycloakRetrieveConfig)
 	.configureRoutes()
 	.configureErrorHandling()
 	.listen();